syfthub-api/pyproject.toml
[project]
name = "syfthub-api"
version = "0.1.0"
description = "A framework for building SyftAI Spaces with a FastAPI-like interface."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
    { name = "OpenMined", email = "info@openmined.org" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Typing :: Typed",
]

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0,<3.0",
    "httpx>=0.27.0",
    "nats-py[aiohttp]>=2.7.0",
    # The syfthub-sdk is a local dependency.
    # It will be installed in editable mode.
    "policy-manager",
    # File-based endpoint mode dependencies
    "watchdog>=3.0.0",
    "python-frontmatter>=1.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    # Subprocess isolation dependencies
    "loky>=3.4.0",
    "packaging>=23.0",
]

[tool.uv.sources]
policy-manager = { git = "ssh://git@github.com/OpenMined/syft-policies.git" }

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.3.0",
    "mypy>=1.8",
    "pre-commit>=3.6",
]

[project.urls]
Homepage = "https://github.com/OpenMined/syfthub"
Documentation = "https://docs.syfthub.org"
Repository = "https://github.com/OpenMined/syfthub"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["syfthub_api", "syfthub_sdk", "policy_manager"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "syfthub_sdk.*",
    "policy_manager.*",
    "uvicorn.*",
    "watchdog.*",
    "frontmatter.*",
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/syfthub_api"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
